+++ src/FluxFlow.Components.Mqtt/Timing/IMqttClock.cs
@@ -1,0 +1,6 @@
+namespace FluxFlow.Components.Mqtt.Timing;
+
+public interface IMqttClock
+{
+    DateTimeOffset UtcNow { get; }
+}